Change header homepage links to use {% url %}, move seach view helpers to helper file
