+++ core/templates/core/privacy_subpage.html
@@ -5,7 +5,7 @@
     <div class="breadcrumbs body-text">
       <ol>
         <li>
-          <a href="/">great.gov.uk</a>
+          <a href="{% url 'landing-page' %}">great.gov.uk</a>
         </li>
         <li>
           <a href="{% url 'privacy-and-cookies' %}">Privacy and cookies</a>
