+++ core/templates/core/search.html
@@ -16,7 +16,7 @@
     <div class="breadcrumbs body-text">
       <ol>
         <li>
-          <a href="/">great.gov.uk</a>
+          <a href="{% url 'landing-page' %}">great.gov.uk</a>
         </li>
         <li aria-current="page">Search</li>
       </ol>
